// src/constants/enums/socialSettings.ts
export enum SocialSettingsForm {
  id = "account__setting_from",

  autocomplete_inputs = "off",
  input_size = "small",
  button_size = "large",
  icon_size = "large",

  facebook_input_id = "facebook_input_id",
  facebook_input_name = "facebook_input",
  facebook_input_label = "",
  facebook_input_role = "facebook_input",
  facebook_input_arial_label = "facebook-input",
  facebook_input_placeholder = "Facebook username",
  facebook_input_type = "text",
  facebook_url = "facebook.com/",

  github_input_id = "github_input_id",
  github_input_name = "github_input",
  github_input_label = "",
  github_input_role = "github_input",
  github_input_arial_label = "github-input",
  github_input_placeholder = "Github username",
  github_input_type = "text",
  github_url = "github.com/",

  twitter_input_id = "twitter_input_id",
  twitter_input_name = "twitter_input",
  twitter_input_label = "",
  twitter_input_role = "twitter_input",
  twitter_input_arial_label = "twitter-input",
  twitter_input_placeholder = "Twitter username",
  twitter_input_type = "text",
  twitter_url = "twitter.com/",

  linkedin_input_id = "linkedin_input_id",
  linkedin_input_name = "linkedin_input",
  linkedin_input_label = "",
  linkedin_input_role = "linkedin_input",
  linkedin_input_arial_label = "linkedin-input",
  linkedin_input_placeholder = "LinkedIn username",
  linkedin_input_type = "text",
  linkedin_url = "linkedin.com/in/",

  instagram_input_id = "instagram_input_id",
  instagram_input_name = "instagram_input",
  instagram_input_label = "",
  instagram_input_role = "instagram_input",
  instagram_input_arial_label = "instagram-input",
  instagram_input_placeholder = "Instagram username",
  instagram_input_type = "text",
  instagram_url = "instagram.com/",

  password_input_id = "password_input_id",
  password_input_name = "password_input",
  password_input_label = "",
  password_input_role = "password-input",
  password_input_arial_label = "password-input",
  password_input_placeholder = "Your password",
  password_input_type_hidden = "password",
  password_input_type_show = "text",
}